requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-widget-tweaks==1.4.8
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
